The reducer now also returns a percentage of correct guesses
