+++ src/features/game/answerReducer.js
@@ -1,30 +1,43 @@
 import { SET_ANSWERED_BREED } from './actions/setAnsweredBreed';
-import { SET_PERCENTAGE } from './actions/setPercentage'
 /**
- * Reducer that stores a list of already answered breeds
+ * Reducer that stores a list of already answered breeds and calculates the percentage of correct guesses in %
  */
 
- const initialState = {
-     answers: [],
-     score: 0
+const initialState = {
+    answers: [],
+    score: 0
 }
 export default (state = initialState, action = {}) => {
     switch (action.type) {
         case SET_ANSWERED_BREED:
-            return {
+            // Construct new state
+            const newState = {
                 ...state,
                 answers: [...state.answers, action.payload]
-            }
-            /*
-            state.answers = [
-                ...state.answers,
-                action.payload
-            ]*/
-        case SET_PERCENTAGE:
+            };
+            // Answers retrieved
+            const percentage = getPercentage(newState.answers);
+            // Return the new state
             return {
-                ...state, score: action.payload / state.answers.length * 100
+                answers: newState.answers,
+                score: percentage / newState.answers.length * 100
             }
         default:
             return state
     }
+}
+
+function getPercentage(arrayOfAnswers) {
+    if (!arrayOfAnswers.length) {
+        return 0
+    }
+    const correctAnswers = arrayOfAnswers.reduce((count, currentAnswer) => {
+        if (currentAnswer.correct) {
+            count++
+        }
+        return count
+    }, 0)
+
+    return correctAnswers
+    //correctAnswers/arrayOfAnswers.length * 100
 }
